+++ components/closet/components/ClothesContainer.tsx
@@ -4,42 +4,20 @@ import { ClothesBox } from 'components/common';
 import { useEffect, useState } from 'react';
 import { productType } from 'types/editPage/product.type';
 type Props = {
-  category: string;
+  clothesData: productType[] | undefined;
 };
-export const ClothesContainer = ({ category }: Props) => {
-  const [clothesData, setClothesData] = useState<Array<productType>>();
-  const getClothes = async () => {
-    await productApi.getFilter({
-        params : {
-          categoryId: category,
-      },
-    }
-      ).then((res) => {
-      console.log(res.data);
-      setClothesData(res.data);
-    }).catch(err => console.log(err));
-    
-    // 옷 한벌 조회
-    // const data = await productApi.getProduct('cl5tir1fc00655gwkgxm6023k');
-    // console.log(data);
-  };
-  useEffect(() => {
-    getClothes();
-  }, []);
-
+export const ClothesContainer = ({ clothesData }: Props) => {
   return (
     <ItemContainer>
-      {clothesData && clothesData.map(
-        (data:productType) =>
-          data.categoryId === "halfT" && (
-            <ClothesBox
-              name={data.name}
-              url={data.imageUrl}
-              key={data.id}
-              type="closet"
-            />
-          ),
-      )}
+      {clothesData &&
+        clothesData.map((data) => (
+          <ClothesBox
+            name={data.name}
+            url={data.imageUrl}
+            key={data.id}
+            type="closet"
+          />
+        ))}
     </ItemContainer>
   );
 };
